pyproject.toml
[project]
name = "sca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "requests>=2.32.5",
]

[tool.basedpyright]
typeCheckingMode = "basic"

[dependency-groups]
def = [
    "djangorestframework-types>=0.9.0",
    "ipython>=9.8.0",
]
dev = [
    "django-types>=0.22.0",
]
